Move delete all vcs
