+++ packages/site/src/pages/cards/DeleteAllVCs.tsx
@@ -1,0 +1,75 @@
+/* eslint-disable no-alert */
+import { IDataManagerClearResult } from '@tuum-tech/identity-snap/src/veramo/plugins/verfiable-creds-manager';
+import { FC, useContext } from 'react';
+import { Card, SendHelloButton } from '../../components';
+import {
+  MetamaskActions,
+  MetaMaskContext,
+} from '../../contexts/MetamaskContext';
+import { VcContext } from '../../contexts/VcContext';
+import {
+  deleteAllVCs,
+  getCurrentNetwork,
+  shouldDisplayReconnectButton,
+} from '../../utils';
+import { validHederaChainID } from '../../utils/hedera';
+
+type Props = {
+  currentChainId: string;
+  setCurrentChainId: React.Dispatch<React.SetStateAction<string>>;
+  hederaAccountConnected: boolean;
+};
+
+const DeleteAllVCs: FC<Props> = ({
+  currentChainId,
+  setCurrentChainId,
+  hederaAccountConnected,
+}) => {
+  const { setVcId, setVcIdsToBeRemoved } = useContext(VcContext);
+  const [state, dispatch] = useContext(MetaMaskContext);
+
+  const handleDeleteAllVCsClick = async () => {
+    try {
+      setCurrentChainId(await getCurrentNetwork());
+      const options = {
+        // If you want to remove the VCs from multiple stores, you can pass an array like so:
+        // store: ['snap', 'googleDrive'],
+        store: 'snap',
+      };
+      const isRemoved = (await deleteAllVCs(
+        options,
+      )) as IDataManagerClearResult[];
+      console.log(`Remove VC Result: ${JSON.stringify(isRemoved, null, 4)}`);
+      setVcId('');
+      setVcIdsToBeRemoved('');
+    } catch (e) {
+      console.error(e);
+      dispatch({ type: MetamaskActions.SetError, payload: e });
+    }
+  };
+
+  return (validHederaChainID(currentChainId) && hederaAccountConnected) ||
+    (!validHederaChainID(currentChainId) && !hederaAccountConnected) ? (
+    <Card
+      content={{
+        title: 'deleteAllVCs',
+        description: 'Delete all the VCs from the snap',
+        button: (
+          <SendHelloButton
+            buttonText="Delete all VCs"
+            onClick={handleDeleteAllVCsClick}
+            disabled={!state.installedSnap}
+          />
+        ),
+      }}
+      disabled={!state.installedSnap}
+      fullWidth={
+        state.isFlask &&
+        Boolean(state.installedSnap) &&
+        !shouldDisplayReconnectButton(state.installedSnap)
+      }
+    />
+  ) : null;
+};
+
+export default DeleteAllVCs;
